+++ lib-unlimited/src/main/java/fr/tduf/libunlimited/high/files/db/patcher/domain/DatabasePatchProperties.java
@@ -194,9 +194,8 @@ public class DatabasePatchProperties extends PatchProperties {
         return retrieve(format(PLACEHOLDER_NAME_FMT_CUSTOM_CAM, cameraView.getPropertySuffix()));
     }
 
-    // TODO set to Optional<Integer>
-    public Optional<Long> getCameraIdentifier() {
-        return retrieve(PLACEHOLDER_NAME_CAMERA).map(Long::valueOf);
+    public Optional<Integer> getCameraIdentifier() {
+        return retrieve(PLACEHOLDER_NAME_CAMERA).map(Integer::valueOf);
     }
 
     public Optional<String> getExteriorMainColorId(int exteriorSet) {
